fixing indicies for sorter, use double

+++ sorter/sorter.cpp
@@ -9,10 +9,10 @@
 #include <memory>
 
 template <typename T>
-std::vector<size_t> sort_indexes(const std::vector<T> &v) {
+std::vector<double> sort_indexes(const std::vector<T> &v) {
 
   // initialize original index locations
-  std::vector<size_t> idx(v.size());
+  std::vector<double> idx(v.size());
   std::iota(idx.begin(), idx.end(), 0);
 
   // sort indexes based on comparing values in v
@@ -20,7 +20,7 @@ std::vector<size_t> sort_indexes(const std::vector<T> &v) {
   // to avoid unnecessary index re-orderings
   // when v contains elements of equal values
   std::stable_sort(idx.begin(), idx.end(),
-       [&v](size_t i1, size_t i2) {return v[i1] < v[i2];});
+       [&v](double i1, double i2) {return v[i1] < v[i2];});
 
   return idx;
 }
@@ -61,7 +61,7 @@ int main(int argc, char** argv)
     in_tt->SetBranchAddress("particle_id",        &m_particle_id);
 
     std::cout << "building entries..\n";
-    std::vector<size_t> indices;
+    std::vector<double> indices;
     for(int i=0; i<in_tt->GetEntries(); ++i){
         in_tt->GetEntry(i);
         indices.push_back(m_hit_start_t_);
